extract all other overlays, move application mode into separate file
