+++ application_mode.py
@@ -1,0 +1,3 @@
+from enum import Enum
+
+application_mode = Enum('application_mode', ['PLAY', 'LEARN_KEY_POINTS', 'LEARN_POINT_HISTORY'])
